Fix an edge case with Jira (#3)
Accounts for a case where Jira's RelayState won't begin with a slash. This happened to me a few minutes ago. Normally, users are redirected to the login page after their session expires. However, sometimes the page will instead show a prompt at the top right to log in again. It turns out that if you click that (at least if your current page is /secure/rapidboard.jspa?...), the RelayState's value won't begin with a slash character. I had been depending on the slash character always being present, but it seems that in this case, it is omitted.

As a side note, sometimes the login prompt doesn't appear and the page instantly boots you to the login page without even setting RelayState to the path you were on. This has been documented in #4.