Add @run-at document-start to improve load time (#9)
* Add @run-at document-start to improve load time

Added @run-at document-start to the metadata to improve load time. Before this, you'd have to wait for the Jira/Confluence landing page to load before you are taken where back to where you were. Now, you're redirected to where you want to go as soon as the browser reaches the landing page.

* Increment version to 0.8